New version with one head
One head predicting on Positive affect18.

Change is concatenating transformer output rather than adding it to the initial input.

Better performance on the positive affect18 by only focusing on one head. The performance on the validation data got up to 55% and on the training data got up to 67%. I didn't test on the test data.